requirements.build.txt
setuptools==75.8.0
fbs @ git+https://github.com/AllYarnsAreBeautiful/fbs@1.2.1-pyqt6
PySide6==6.6.2
PyInstaller==6.4.0
Pillow==10
pyserial==3.5
sliplib==0.6.2
bitarray==2.9.2
simpleaudio==1.0.4
numpy==1.26.4
charset-normalizer<3.0.0
semver==3.0.2